Add recursion Practice Script file, update opt1
The recursion file introduces a binary search tree implementation,
and uses that for some practice.  More practice problems could
come from this, but it may be introducing too much at once.

opt1.py needs scope and exception problems, I have an idea for
both but they are a work in progress.